bid overlay broken down into components
