+++ packages/desktop/src/components/bids/components/bid/index.js
@@ -1,0 +1,3 @@
+import Bid from './bid';
+
+export default Bid;
