+++ packages/desktop/src/components/bids/components/bids/index.js
@@ -1,0 +1,3 @@
+import Bids from './bids';
+
+export default Bids;
