+++ packages/desktop/src/components/bids/components/newBid/index.js
@@ -1,0 +1,3 @@
+import NewBid from './newBid';
+
+export default NewBid;
